+++ soroban-spec/src/rust/types.rs
